[#21] ref: src/cleanup: separate files
